% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_dataset.R
\name{baseline_dataset}
\alias{baseline_dataset}
\title{Load baseline simulated dataset for trying EpiFusion}
\usage{
baseline_dataset()
}
\description{
This function loads a simple simulated outbreak dataset (a phylogenetic tree and weekly case incidence) to your global environment
}
