% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EpiFusionUtilities-package.R
\docType{package}
\name{EpiFusionUtilities}
\alias{EpiFusionUtilities-package}
\alias{EpiFusionUtilities}
\title{EpiFusionUtilities: Tools for Implementing EpiFusion Joint Inference Models}
\description{
The EpiFusionUtilities package provides a suite of functions for preparing data
to run joint inference models using the EpiFusion particle filtering infrastructure.
It is also possible to run EpiFusion from within you R session using the package, and
parse the results.
}
\section{Main Functions}{

\itemize{
\item \code{prepare_epifusion_tree}: For preparing time rooted phylogenetic tree(s).
\item \code{generate_epifusion_XML}: For generating EpiFusion XML input files.
\item \code{run_epifusion}: For running EpiFusion within R.
\item \code{load_raw_epifusion}: For loading raw results of EpiFusion analysis.
\item \code{extract_posterior_epifusion}: For discarding MCMC chain burnin and extracting parameter posteriors.
}
}

\section{Installation}{

You can install the development version from GitHub with:

\if{html}{\out{<div class="sourceCode">}}\preformatted{devtools::install_github("https://github.com/ciarajudge/EpiFusionUtilities")
}\if{html}{\out{</div>}}
}

\section{Author(s)}{

Ciara Judge \href{mailto:ciara.judge@lshtm.ac.uk}{ciara.judge@lshtm.ac.uk}
}

\section{Maintainer}{

Ciara Judge \href{mailto:ciara.judge@lshtm.ac.uk}{ciara.judge@lshtm.ac.uk}
}

\examples{
# Example usage
library(EpiFusionUtilities)
run_epifusion("path/to/EpiFusion/XML/file")
}
\author{
\strong{Maintainer}: Ciara Judge \email{ciara.judge@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-7923-9898}{ORCID})

}
\keyword{internal}
