% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_XML_chunk.R
\name{generate_XML_chunk}
\alias{generate_XML_chunk}
\title{Generate an XML chunk from a list}
\usage{
generate_XML_chunk(plist, node)
}
\arguments{
\item{plist}{...}

\item{node}{...}
}
\description{
Th
}
