% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_beta_trajectories.R
\name{load_beta_trajectories}
\alias{load_beta_trajectories}
\title{Load beta trajectories from EpiFusion folder into an list of dataframes
This function loads the beta trajectories of an EpiFusion analysis into a list of dataframes, with one dataframe for each chain in the analysis. The data frames are structured with ncol = number of days in the analysis, and nrow = number of MCMC samples.}
\usage{
load_beta_trajectories(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
a list of beta trajectory dataframes
}
\description{
Load beta trajectories from EpiFusion folder into an list of dataframes
This function loads the beta trajectories of an EpiFusion analysis into a list of dataframes, with one dataframe for each chain in the analysis. The data frames are structured with ncol = number of days in the analysis, and nrow = number of MCMC samples.
}
