% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cumulativeinfection_trajectories.R
\name{load_cumulativeinfection_trajectories}
\alias{load_cumulativeinfection_trajectories}
\title{Load cumulative infections trajectories from EpiFusion output folder into a list of dataframes}
\usage{
load_cumulativeinfection_trajectories(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
a list of cumulative trajectory dataframes
}
\description{
This function loads the cumulative infection trajectories of an EpiFusion analysis into a list of dataframes, with one dataframe for each chain in the analysis. The data frames are structured with ncol = number of days in the analysis, and nrow = number of MCMC samples.
}
