% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epifusion.R
\name{load_epifusion}
\alias{load_epifusion}
\title{Load results from EpiFusion folder into an R object}
\usage{
load_epifusion(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
Temperature in Celsius
}
\description{
Load results from EpiFusion folder into an R object
}
