% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_likelihoods.R
\name{load_likelihoods}
\alias{load_likelihoods}
\title{Load results from EpiFusion folder into an R object}
\usage{
load_likelihoods(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
a list of R compatible EpiFusion output objects
}
\description{
This function loads the result of an EpiFusion folder into an R list, with the key characteristics and results of the analysis included.
}
