% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameter_samples.R
\name{load_parameter_samples}
\alias{load_parameter_samples}
\title{Load results from EpiFusion folder into an R object
This function loads the result of an EpiFusion folder into an R list, with the key characteristics and results of the analysis included.}
\usage{
load_parameter_samples(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
a list of R compatible EpiFusion output objects
}
\description{
Load results from EpiFusion folder into an R object
This function loads the result of an EpiFusion folder into an R list, with the key characteristics and results of the analysis included.
}
