% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chainwise_trajectories.R
\name{plot_chainwise_trajectories}
\alias{plot_chainwise_trajectories}
\title{Plot the trajectory posteriors from a raw EpiFusion object separated by chain}
\usage{
plot_chainwise_trajectories(raw_epifusion, burn_in, type = NA)
}
\arguments{
\item{raw_epifusion}{raw EpiFusion object}

\item{burn_in}{proportion (as a decimal) of each chain to discard as burn in}

\item{type}{(optional) "infection" "rt" or "cumulativeinfection"}
}
\description{
Plots the trajectory posteriors from a raw EpiFusion object, separated by chain to allow assessment of convergence. You can specify the trajectory type (infection, Rt, cumulative infection) using the \code{type} parameter, or the function will plot each type by default. You may specify a proportion of each chain to discard as burn-in with the \code{burn_in} parameter.
}
