% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likelihood_trace.R
\name{plot_likelihood_trace}
\alias{plot_likelihood_trace}
\title{Plots the likelihood trace of the chains of an EpiFusion analysis}
\usage{
plot_likelihood_trace(raw_epifusion)
}
\arguments{
\item{raw_epifusion}{raw epifusion object}
}
\description{
This function plots the likelihood trace of the chains of an EpiFusion analysis to assess convergence and help to select the burn-in proportion.
}
