% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trajectories.R
\name{plot_trajectories}
\alias{plot_trajectories}
\title{Plot the trajectory posteriors from a parsed EpiFusion trajectory table}
\usage{
plot_trajectories(trajectory_table, type = NA, plot_colours = NA)
}
\arguments{
\item{trajectory_table}{a trajectory table extracted from an EpiFusion posterior object}

\item{type}{(optional) "infection" "rt" or "cumulativeinfection"}

\item{plot_colours}{(optional) a vector of colours to be used in the plots}
}
\description{
T
}
