% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_epifusion.R
\name{run_epifusion}
\alias{run_epifusion}
\title{Run EpiFusion from R
This function runs EpiFusion from R, assuming you have java installed.}
\usage{
run_epifusion(xml_filepath, output_folder_name)
}
\arguments{
\item{xml_filepath}{path to the xml input file}

\item{output_folder_name}{the name of the folder you'd like created for the output}
}
\description{
Run EpiFusion from R
This function runs EpiFusion from R, assuming you have java installed.
}
