% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_dataset.R
\name{sampling_dataset}
\alias{sampling_dataset}
\title{Load step-change in sampling simulated dataset for trying EpiFusion}
\usage{
sampling_dataset()
}
\description{
This function loads a simple simulated outbreak dataset (a phylogenetic tree and weekly case incidence) to your global environment. In this dataset, sampling of both case incidence and genomic sequence data is initially low, followed by a sharp increase.
}
