% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_posterior_epifusion.R
\name{extract_posterior_epifusion}
\alias{extract_posterior_epifusion}
\title{Load results from EpiFusion folder into an R object
This function loads the result of an EpiFusion folder into an R list, with the key characteristics and results of the analysis included.}
\usage{
extract_posterior_epifusion(raw_epifusion, burn_in)
}
\arguments{
\item{raw_epifusion}{filepath to folder}

\item{burn_in}{proportion (as a decimal) of chains to discard as burn in}
}
\value{
a list of R compatible EpiFusion output objects
}
\description{
Load results from EpiFusion folder into an R object
This function loads the result of an EpiFusion folder into an R list, with the key characteristics and results of the analysis included.
}
