% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_acceptance.R
\name{load_acceptance}
\alias{load_acceptance}
\title{Load acceptance rates from EpiFusion folder into an R object}
\usage{
load_acceptance(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
a list of vectors of acceptance rates of each chain
}
\description{
This function loads the acceptance rates for each chain of an EpiFusion analysis into a list of vectors.
}
