% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_likelihoods.R
\name{load_likelihoods}
\alias{load_likelihoods}
\title{Load likelihood trace from each chain of an EpiFusion analysis}
\usage{
load_likelihoods(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
a list of numeric vectors of likelihoods from an EpiFusion analysis
}
\description{
This function loads the likelihood traces of each chain of an EpiFusion analysis into a list of vectors, with one vector for each chain
}
