% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameter_samples.R
\name{load_parameter_samples}
\alias{load_parameter_samples}
\title{Load parameter samples from EpiFusion analysis}
\usage{
load_parameter_samples(folderpath)
}
\arguments{
\item{folderpath}{filepath to folder}
}
\value{
a list of R compatible EpiFusion output objects
}
\description{
This function loads the parameter samples from an EpiFusion analysis into an R list, with one data frame per chain, with one column per parameter and one row per MCMC sample
}
