% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameter_trace.R
\name{plot_parameter_trace}
\alias{plot_parameter_trace}
\title{Plot the parameter traces from a raw EpiFusion object}
\usage{
plot_parameter_trace(raw_epifusion)
}
\arguments{
\item{raw_epifusion}{raw epifusion object}
}
\description{
T
}
