% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_epifusion_tree.R
\name{prepare_epifusion_tree}
\alias{prepare_epifusion_tree}
\title{Prepare a tree (in S3 Phylo Object form) for EpiFusion}
\usage{
prepare_epifusion_tree(tree)
}
\arguments{
\item{tree}{phylogenetic tree (in S3 Phylo Object format)}
}
\value{
a tree (in S3 Phylo Object format)
}
\description{
This function prepares a tree, given as an S3 phylo object, for EpiFusion analysis by adding 'time during outbreak' to the node and leaf labels.
}
