% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory_table.R
\name{trajectory_table}
\alias{trajectory_table}
\title{Extract a table with the mean and HPD intervals of infection and Rt trajectories from an EpiFusion posterior R object
This function extracts a table}
\usage{
trajectory_table(epifusion_posterior, start_date = NA)
}
\arguments{
\item{epifusion_posterior}{an extracted EpiFusion posterior object made with the extract_posterior_epifusion() function}

\item{start_date}{(optional) date you'd like the time series to count from, as a string with format YYYY-MM-DD}
}
\value{
a table with a column for Time and a column for Mean, Upper and Lower 95\%, 88\% and 66\% HPDs for inferred Infections and Rt
}
\description{
Extract a table with the mean and HPD intervals of infection and Rt trajectories from an EpiFusion posterior R object
This function extracts a table
}
